Update README.md
Here is a revised version. I am still working through this. 

For the  Socure Predictive Docv node configuration settings: 
- Is the endpoint URL `https://sandbox.socure.com/api/3.0/EmailAuthScore`? I think the Web SDK uses the Document Request endpoint: https://service.socure.com/api/3.0/documents/request
- What is the `isDocVRequired` parameter? This toggles the node on and off? 

Also, the hyperlink does not work in this text: 
"The code in this repository has binary dependencies that live in the ForgeRock Maven repository. Maven can be configured to authenticate to this repository by following the steps in the [ForgeRock Knowledge Base Article](https://backstage.forgerock.com/knowledge/kb/article/a74096897)."